+++ api/src/main/java/com/flipkart/poseidon/internal/DataSourceCallable.java
@@ -20,6 +20,7 @@ import com.flipkart.hydra.composer.Composer;
 import com.flipkart.hydra.composer.exception.ComposerEvaluationException;
 import com.flipkart.hydra.task.entities.WrapperCallable;
 import com.flipkart.hydra.task.exception.BadCallableException;
+import com.flipkart.poseidon.datasources.DataSourceRequest;
 import com.flipkart.poseidon.legoset.PoseidonLegoSet;
 import com.google.common.util.concurrent.ListenableFuture;
 import flipkart.lego.api.entities.DataSource;